Project/app/requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
branca==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
executing==2.2.0
fastapi==0.115.6
Flask==3.1.0
folium==0.19.4
fonttools==4.55.6
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mercantile==1.2.1
networkx==3.4.2
numpy==2.2.2
osmnx==2.0.1
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
rasterio==1.4.3
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
style==1.1.0
threadpoolctl==3.5.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
update==0.0.1
urllib3==2.3.0
utils==1.0.2
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
xyzservices==2025.1.0
